Vue player viewer and relevant buttons
