+++ website/src/requests/GetShopCatalogRequest.js
@@ -1,0 +1,8 @@
+export async function getShopCatalogRequest(server) {
+    const response = await fetch(server + 'get_shop_catalog?version=1')
+    const responseBody = await response.text();
+    if(response.status != 200) {
+        return false
+    }
+    return JSON.parse(responseBody)
+}
